# src/backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Specvital API
  description: Test specification analysis service for GitHub repositories
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/auth/login:
    get:
      operationId: authLogin
      summary: Initiate GitHub OAuth login
      description: Redirects to GitHub OAuth authorization page
      responses:
        "200":
          description: OAuth authorization URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/auth/callback:
    get:
      operationId: authCallback
      summary: GitHub OAuth callback
      description: Handles OAuth callback from GitHub and sets authentication cookie
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: OAuth authorization code from GitHub
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: OAuth state for CSRF protection
      responses:
        "302":
          description: Redirect to frontend after successful authentication
          headers:
            Location:
              schema:
                type: string
              description: Frontend URL to redirect to
            Set-Cookie:
              schema:
                type: string
              description: HTTP-only authentication cookie
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/auth/logout:
    post:
      operationId: authLogout
      summary: Logout and clear authentication
      description: Clears authentication cookies and revokes refresh token in database
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/auth/refresh:
    post:
      operationId: authRefresh
      summary: Refresh authentication tokens
      description: |
        Exchanges a valid refresh token for a new access/refresh token pair.
        The old refresh token is invalidated (rotation).
        If a revoked token is used, the entire token family is invalidated (reuse detection).
      responses:
        "200":
          description: Tokens refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
              description: New access and refresh tokens as HTTP-only cookies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/auth/dev-login:
    post:
      operationId: authDevLogin
      summary: Development-only test login
      description: |
        Creates or retrieves a test user and issues authentication tokens.
        Only available when ENV != "production".
        WARNING: This endpoint is for development and E2E testing only.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevLoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Authentication cookies (auth_token, refresh_token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevLoginResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT token stored in HTTP-only cookie
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    Forbidden:
      description: Access forbidden
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
...
    ProblemDetail:
      type: object
      required:
        - status
        - title
        - detail
      properties:
        type:
          type: string
          format: uri
          default: about:blank
          description: URI identifying the problem type
        title:
          type: string
          description: Human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
        instance:
          type: string
          description: URI reference identifying the specific occurrence
        rateLimit:
          $ref: "#/components/schemas/RateLimitInfo"

    # Rate Limit Info
    RateLimitInfo:
      type: object
      required:
        - limit
        - remaining
        - resetAt
      properties:
        limit:
          type: integer
          description: Maximum requests allowed per time window
        remaining:
          type: integer
          description: Remaining requests in current window
        resetAt:
          type: integer
          format: int64
          description: Unix timestamp when the rate limit resets

    # Auth Schemas
    LoginResponse:
      type: object
      required:
        - authUrl
      properties:
        authUrl:
          type: string
          format: uri
          description: GitHub OAuth authorization URL to redirect user
    UserInfo:
      type: object
      required:
        - id
        - login
        - avatarUrl
      properties:
        id:
          type: string
          description: Internal user ID
        login:
          type: string
          description: GitHub username
        avatarUrl:
          type: string
          format: uri
          description: GitHub avatar URL
        name:
          type: string
          description: GitHub display name (optional)
    LogoutResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Logout operation result
    RefreshResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Token refresh operation result
    DevLoginRequest:
      type: object
      properties:
        userId:
          type: string
          description: Optional user ID to login as (uses default test user if not provided)
        username:
          type: string
          description: Optional username for test user (default is "test-user")
    DevLoginResponse:
      type: object
      required:
        - success
        - user
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/UserInfo"

    # Repository List Responses
...